requirements.txt
aiofiles==23.2.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
piccolo==1.27.1
platformdirs==4.3.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
targ==0.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
